chore: export factory from border
